♻ Refactor transaction generator
